backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.7
backports.tarfile==1.2.0
beartype==0.22.9
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
croniter==6.0.0
cryptography==46.0.4
cyclopts==4.5.1
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.128.4
fastmcp==2.14.5
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.2.0
gradio_client==2.0.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lupa==2.6
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.1
numpy==2.2.6
openapi-pydantic==0.5.1
opentelemetry-api==1.39.1
packaging==26.0
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.24.1
propcache==0.4.1
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.17.5
Pygments==2.19.2
PyJWT==2.11.0
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
referencing==0.36.2
requests==2.32.5
rich==14.3.2
rich-rst==1.3.2
rpds-py==0.30.0
SecretStorage==3.5.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sse-starlette==3.2.0
starlette==0.52.1
taskgroup==0.2.2
tomli==2.4.0
tqdm==4.67.3
trimesh==4.11.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==16.0
yarl==1.22.0
zipp==3.23.0
